+++ testing/unit_tests.py
@@ -1,0 +1,11 @@
+import unittest
+
+from pyFireEyeAlert import *
+
+class PrimesTestCase(unittest.TestCase):
+    """Tests for `primes.py`."""
+
+
+
+if __name__ == '__main__':
+    unittest.main()